Move timestamp function tests to TimestampFunctionsSuite
